Implement paywall tester scenario and generate baseline profiles
