+++ baselineprofile/build.gradle.kts
@@ -59,7 +59,7 @@ android {
         create<ManagedVirtualDevice>("pixel6Api34") {
             device = "Pixel 6"
             apiLevel = 34
-            systemImageSource = "google"
+            systemImageSource = "aosp"
         }
     }
 }
